requirements.txt
confluent_kafka==2.3.0
Flask==3.0.0
Mesa==2.1.2
numpy==1.26.2
python-dotenv==1.0.0
